Mode LocalFeedItem to its own file
