+++ EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -1,20 +1,5 @@
 import Foundation
 
-public struct LocalFeedItem: Equatable {
-    public let id: UUID
-    public let description: String?
-    public let location: String?
-    public let imageURL: URL
-
-    public init (id: UUID, description: String?, location: String?, imageURL: URL) {
-        self.id = id
-        self.description = description
-        self.location = location
-        self.imageURL = imageURL
-    }
-}
-
-
 public final class LocalFeedLoader {
     private let store: FeedStore
     private let currentDate: () -> Date
